fix streaming cache robustness
