+++ crates/once-core/src/action.rs
@@ -1,0 +1,74 @@
+use std::collections::BTreeMap;
+
+use once_cas::Digest;
+use serde::{Deserialize, Serialize};
+
+use crate::{ResourceRequest, WorkspacePath};
+
+/// Domain-separation prefix for action digests. Bump the version when
+/// the canonical encoding (or the [`Action`] schema) changes in a way
+/// that should invalidate the cache.
+pub(crate) const ACTION_DIGEST_DOMAIN: &[u8] = b"once.action.v3\0";
+
+/// All actions Once can execute.
+///
+/// The wire format of this enum is part of the action digest (see
+/// `ACTION_DIGEST_DOMAIN`). Field additions, renames, or reorderings
+/// that affect the JSON encoding require a digest version bump.
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(tag = "kind", rename_all = "snake_case")]
+pub enum Action {
+    RunCommand {
+        argv: Vec<String>,
+        #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
+        env: BTreeMap<String, String>,
+        #[serde(default, skip_serializing_if = "Option::is_none")]
+        cwd: Option<WorkspacePath>,
+        #[serde(default, skip_serializing_if = "Option::is_none")]
+        input_digest: Option<Digest>,
+        /// Workspace-relative paths the action promises to produce. The
+        /// runner stores each one in the CAS after a fresh execution
+        /// and restores it from the CAS on a cache hit. An empty list
+        /// means the action has no declared outputs (only stdout/stderr
+        /// are cached); cache hits then provide nothing on disk.
+        #[serde(default, skip_serializing_if = "Vec::is_empty")]
+        outputs: Vec<WorkspacePath>,
+        #[serde(default, skip_serializing_if = "ResourceRequest::is_default")]
+        resources: ResourceRequest,
+        /// Per-action timeout in milliseconds. None = no timeout.
+        #[serde(default, skip_serializing_if = "Option::is_none")]
+        timeout_ms: Option<u64>,
+        /// Optional compute provider for remote execution. This is
+        /// part of the action key so local and remote runs never share
+        /// a cache slot by accident.
+        #[serde(default, skip_serializing_if = "Option::is_none")]
+        remote: Option<RemoteExecution>,
+    },
+}
+
+impl Action {
+    /// Canonical, content-addressed key for this action.
+    ///
+    /// The key is `BLAKE3(domain || canonical_json(self))`. Bumping the
+    /// domain partitions old and new cache entries cleanly instead of
+    /// silently colliding.
+    pub fn digest(&self) -> Digest {
+        let body = serde_json::to_vec(self).expect("Action is serializable");
+        let mut buf = Vec::with_capacity(ACTION_DIGEST_DOMAIN.len() + body.len());
+        buf.extend_from_slice(ACTION_DIGEST_DOMAIN);
+        buf.extend_from_slice(&body);
+        Digest::of_bytes(&buf)
+    }
+
+    pub fn resource_request(&self) -> &ResourceRequest {
+        match self {
+            Action::RunCommand { resources, .. } => resources,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub struct RemoteExecution {
+    pub provider: String,
+}
