+++ crates/once-core/src/execute.rs
@@ -1,0 +1,67 @@
+use std::path::Path;
+
+use once_cas::{ActionResult, CacheProvider};
+
+use crate::{local, outputs, remote, Action, Result};
+
+pub(crate) async fn run(
+    action: &Action,
+    workspace_root: &Path,
+    cache: &CacheProvider,
+    stream_to_parent: bool,
+) -> Result<ActionResult> {
+    match action {
+        Action::RunCommand {
+            argv,
+            env,
+            cwd,
+            input_digest: _,
+            outputs,
+            resources: _,
+            timeout_ms,
+            remote,
+        } => {
+            let mut result = match (remote, stream_to_parent) {
+                (Some(remote), _) => {
+                    remote::execute_command(
+                        remote,
+                        argv,
+                        env,
+                        cwd.as_ref(),
+                        *timeout_ms,
+                        workspace_root,
+                        cache,
+                        stream_to_parent,
+                    )
+                    .await?
+                }
+                (None, true) => {
+                    Box::pin(local::execute_command_streaming(
+                        argv,
+                        env,
+                        cwd.as_ref(),
+                        *timeout_ms,
+                        workspace_root,
+                        cache,
+                    ))
+                    .await?
+                }
+                (None, false) => {
+                    local::execute_command(
+                        argv,
+                        env,
+                        cwd.as_ref(),
+                        *timeout_ms,
+                        workspace_root,
+                        cache,
+                    )
+                    .await?
+                }
+            };
+            if result.exit_code == 0 {
+                result.outputs = outputs::capture(outputs, workspace_root, cache).await?;
+            }
+            Ok(result)
+        }
+    }
+}
